add new docker image and remove large file from integration test
