+++ ci/jobs/scripts/integration_tests_configs.py
@@ -61,6 +61,7 @@ IMAGES_ENV = {
     "clickhouse/python-bottle": "DOCKER_PYTHON_BOTTLE_TAG",
     "clickhouse/integration-test-with-unity-catalog": "DOCKER_BASE_WITH_UNITY_CATALOG_TAG",
     "clickhouse/integration-test-with-hms": "DOCKER_BASE_WITH_HMS_TAG",
+    "clickhouse/integration-test-with-paimon-rest-catalog": "DOCKER_BASE_WITH_PAIMON_REST_CATALOG_TAG",
     "clickhouse/mysql_dotnet_client": "DOCKER_MYSQL_DOTNET_CLIENT_TAG",
     "clickhouse/s3-proxy": "DOCKER_S3_PROXY_TAG",
 }
