+++ tests/integration/compose/docker_compose_paimon_rest_catalog.yml
@@ -1,0 +1,39 @@
+services:
+  node:
+    volumes:
+      - paimon-warehouse:/var/lib/clickhouse/user_files/warehouse
+  paimon-rest-bearer:
+    image: clickhouse/integration-test-with-paimon-rest-catalog:${DOCKER_BASE_WITH_PAIMON_REST_CATALOG_TAG:-latest}
+    restart: "no"
+    expose:
+      - 8001
+    volumes:
+      - paimon-warehouse:/var/lib/clickhouse/user_files/warehouse
+    command: ["server", "file:///var/lib/clickhouse/user_files/warehouse/", "bearer", "0.0.0.0", "8001"]
+    healthcheck:
+      test: ["CMD", "bash", "-c", "echo > /dev/tcp/localhost/8001"]
+      interval: 2s
+      timeout: 5s
+      retries: 30
+      start_period: 10s
+    stop_grace_period: 5s
+    cpus: 2
+  paimon-rest-dlf:
+    image: clickhouse/integration-test-with-paimon-rest-catalog:${DOCKER_BASE_WITH_PAIMON_REST_CATALOG_TAG:-latest}
+    restart: "no"
+    expose:
+      - 8002
+    volumes:
+      - paimon-warehouse:/var/lib/clickhouse/user_files/warehouse
+    command: ["server", "file:///var/lib/clickhouse/user_files/warehouse/", "dlf", "0.0.0.0", "8002"]
+    healthcheck:
+      test: ["CMD", "bash", "-c", "echo > /dev/tcp/localhost/8002"]
+      interval: 2s
+      timeout: 5s
+      retries: 30
+      start_period: 10s
+    stop_grace_period: 5s
+    cpus: 2
+
+volumes:
+  paimon-warehouse:
