--- tests/integration/test_paimon_rest_catalog/paimon-rest-catalog/.gitignore
@@ -1,2 +1,0 @@
-target/
-dependency-reduced-pom.xml
