--- tests/integration/test_paimon_rest_catalog/paimon-rest-catalog/src/main/resources/META-INF/services/org.apache.paimon.factories.Factory
@@ -1,44 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.
-# The ASF licenses this file to You under the Apache License, Version 2.0
-# (the "License"); you may not use this file except in compliance with
-# the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-org.apache.paimon.catalog.FileSystemCatalogFactory
-org.apache.paimon.jdbc.JdbcCatalogFactory
-org.apache.paimon.jdbc.JdbcCatalogLockFactory
-org.apache.paimon.mergetree.compact.aggregate.factory.FieldBoolAndAggFactory
-org.apache.paimon.mergetree.compact.aggregate.factory.FieldBoolOrAggFactory
-org.apache.paimon.mergetree.compact.aggregate.factory.FieldCollectAggFactory
-org.apache.paimon.mergetree.compact.aggregate.factory.FieldFirstNonNullValueAggFactory
-org.apache.paimon.mergetree.compact.aggregate.factory.FieldFirstNonNullValueAggLegacyFactory
-org.apache.paimon.mergetree.compact.aggregate.factory.FieldFirstValueAggFactory
-org.apache.paimon.mergetree.compact.aggregate.factory.FieldHllSketchAggFactory
-org.apache.paimon.mergetree.compact.aggregate.factory.FieldLastNonNullValueAggFactory
-org.apache.paimon.mergetree.compact.aggregate.factory.FieldLastValueAggFactory
-org.apache.paimon.mergetree.compact.aggregate.factory.FieldListaggAggFactory
-org.apache.paimon.mergetree.compact.aggregate.factory.FieldMaxAggFactory
-org.apache.paimon.mergetree.compact.aggregate.factory.FieldMergeMapAggFactory
-org.apache.paimon.mergetree.compact.aggregate.factory.FieldMinAggFactory
-org.apache.paimon.mergetree.compact.aggregate.factory.FieldNestedUpdateAggFactory
-org.apache.paimon.mergetree.compact.aggregate.factory.FieldPrimaryKeyAggFactory
-org.apache.paimon.mergetree.compact.aggregate.factory.FieldProductAggFactory
-org.apache.paimon.mergetree.compact.aggregate.factory.FieldRoaringBitmap32AggFactory
-org.apache.paimon.mergetree.compact.aggregate.factory.FieldRoaringBitmap64AggFactory
-org.apache.paimon.mergetree.compact.aggregate.factory.FieldSumAggFactory
-org.apache.paimon.mergetree.compact.aggregate.factory.FieldThetaSketchAggFactory
-org.apache.paimon.rest.RESTCatalogFactory
-org.apache.paimon.iceberg.migrate.IcebergMigrateHadoopMetadataFactory
-org.apache.paimon.rest.auth.BearTokenAuthProviderFactory
-org.apache.paimon.rest.auth.DLFAuthProviderFactory
-org.apache.paimon.rest.auth.DLFLocalFileTokenLoaderFactory
-org.apache.paimon.rest.auth.DLFECSTokenLoaderFactory
